requirements.txt

pyo==1.0.5      # http://ajaxsoundstudio.com/software/pyo/
# pyserial==3.5   # https://pyserial.readthedocs.io/en/latest/index.html

# pip~=22.0.4
pyserial~=3.5
# setuptools~=56.0.0